Extracted tests spring configuration to TestsGeneratorCOnfiguration class
